ebdjango/requirements.txt
asgiref==3.8.1
boto3==1.34.101
botocore==1.34.101
cffi==1.16.0
colorama==0.4.6
cryptography==42.0.7
Django==5.0.4
psycopg2-binary
django-cors-headers==4.3.1
djangorestframework==3.15.1
jmespath==1.0.1
PyJWT==2.8.0
pycparser==2.22
python-dateutil==2.9.0.post0
s3transfer==0.10.1
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
inflection
pyyaml
django-rest-swagger
Pillow